First upload
This responsive site, crafted with HTML, CSS, and JS (for button bact to top) brings a fresh and engaging experience to food enthusiasts. The repository features a responsive header with a logo, navigation, and a compelling call-to-action that optimally adapts to mobile devices.

The elegant hero banner captures the essence of our culinary journey, setting the stage for an immersive experience. The main section seamlessly integrates about us, a mouth-watering menu, customer reviews, a convenient contact form, and captivating blog content—all beautifully displayed in responsive flexbox layouts.

Every element, from the stylish card-based design to the mobile-friendly arrangements, is meticulously crafted to provide a delightful user experience. The display:flex property ensures a consistent and visually appealing presentation across devices.

Experience the ease of navigation and visually appealing content that this repository offers. Don't forget to explore the "Back to Top" button for added convenience. Dive into the world of flavors, and happy coding! 🍽️✨